--- Gatekeeper/gatekeeper/common/models.py
@@ -1,3 +1,0 @@
-import attr
-
-data_model = attr.s(slots=True, auto_attribs=True)
