--- Gatekeeper/gatekeeper/webhooks/twilio_views/message_status.py
@@ -1,10 +1,0 @@
-from django.views.decorators.http import require_POST
-
-from common.response import success_response
-
-from ..models.twilio_webhook_request import TwilioWebhookRequest
-
-
-@require_POST
-def message_status_webhook(request):
-    return success_response()
