+++ Gatekeeper/pyproject.toml
@@ -20,4 +20,4 @@ include_trailing_comma = true
 force_grid_wrap = 0
 use_parentheses = true
 line_length = 88
-known_third_party = "twilio"
+known_third_party = "attrs,twilio"
